DRY out roundtrip config fixtures into `config_fixture` factory
